# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Phantom Mask API",
        "description": "API for managing pharmacies, masks, and purchases",
        "contact": {
            "email": "rouyucc@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/customers/top/{count}": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "統計指定日期區間購買口罩總金額最高的前 X 名使用者",
                "description": "回傳在指定日期範圍內，依照購買總金額排名的前幾名顧客列表。",
                "operationId": "getTopCustomers",
                "parameters": [
                    {
                        "name": "count",
                        "in": "path",
                        "description": "要列出前幾名顧客",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "開始日期 (YYYY-MM-DD)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "結束日期 (YYYY-MM-DD)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "顧客排名列表",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "John Doe"
                                            },
                                            "total_amount": {
                                                "description": "指定日期範圍內的總交易金額",
                                                "type": "number",
                                                "format": "float",
                                                "example": 123.45
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "無效的參數請求",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "無效的參數請求"
                                        },
                                        "errors": {
                                            "properties": {
                                                "start_date": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "開始日期欄位必填"
                                                    }
                                                },
                                                "end_date": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": [
                                                            "結束日期欄位必填",
                                                            "結束日期必須等於或晚於開始日期"
                                                        ]
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}